requirements_release.txt
azureml-sdk
numpy